Add signin page instead of modal
